【paddle.fleet】raise error when using multi-cards in fleet non_distributed mode (#27854)
* raise error if use multi-cards in fleet non_distributed mode; test=develop
